src/app/modules/user/domain/usecases/create.user.ts
import { Either } from 'fp-ts/lib/Either';
import { Failure } from '../errors/failure';
import { IUser } from '../interfaces/user.entity.interface';
import { IUserRepository } from '../interfaces/user.repository.interface';

export class CreateUser {
  userRepository: IUserRepository;

  constructor(userRepository: IUserRepository) {
    this.userRepository = userRepository;
  }

  async call(name: string, email: string): Promise<Either<Failure, IUser>> {
    return await this.userRepository.createUser(name, email);
  }
}
